requirements.txt
Flask
tensorflow
pillow
python-dotenv
